Move templating functionality into its own package from Levant pkg
Closes #171
