+++ template/render.go
@@ -1,11 +1,10 @@
-package levant
+package template
 
 import (
 	"bytes"
 	"fmt"
 	"io/ioutil"
 	"path"
-	"text/template"
 
 	"github.com/jrasell/levant/helper"
 	"github.com/rs/zerolog/log"
@@ -14,12 +13,6 @@ import (
 	nomad "github.com/hashicorp/nomad/api"
 	"github.com/hashicorp/nomad/jobspec"
 	"github.com/hashicorp/terraform/config"
-)
-
-const (
-	terraformVarExtension = ".tf"
-	yamlVarExtension      = ".yaml"
-	ymlVarExtension       = ".yml"
 )
 
 // RenderJob takes in a template and variables performing a render of the
@@ -39,9 +32,9 @@ func RenderJob(templateFile, variableFile string, flagVars *map[string]string) (
 // passed variables file.
 func RenderTemplate(templateFile, variableFile string, flagVars *map[string]string) (tpl *bytes.Buffer, err error) {
 	if variableFile == "" {
-		log.Debug().Msgf("levant/templater: no variable file passed, trying defaults")
+		log.Debug().Msgf("template/render: no variable file passed, trying defaults")
 		if variableFile = helper.GetDefaultVarFile(); variableFile != "" {
-			log.Debug().Msgf("levant/templater: found default variable file, using %s", variableFile)
+			log.Debug().Msgf("template/render: found default variable file, using %s", variableFile)
 		}
 	}
 
@@ -49,7 +42,7 @@ func RenderTemplate(templateFile, variableFile string, flagVars *map[string]stri
 	// correctly rendered.
 	var ext string
 	if ext = path.Ext(variableFile); ext != "" {
-		log.Debug().Msgf("levant/templater: variable file extension %s detected", ext)
+		log.Debug().Msgf("template/render: variable file extension %s detected", ext)
 	}
 
 	src, err := ioutil.ReadFile(templateFile)
@@ -60,7 +53,7 @@ func RenderTemplate(templateFile, variableFile string, flagVars *map[string]stri
 	// If no command line variables are passed; log this as DEBUG to provide much
 	// greater feedback.
 	if len(*flagVars) == 0 {
-		log.Debug().Msgf("levant/templater: no command line variables passed")
+		log.Debug().Msgf("template/render: no command line variables passed")
 	}
 
 	switch ext {
@@ -73,7 +66,7 @@ func RenderTemplate(templateFile, variableFile string, flagVars *map[string]stri
 
 	case "":
 		// No variables file passed; render using any passed CLI variables.
-		log.Debug().Msgf("levant/templater: variable file not passed")
+		log.Debug().Msgf("template/render: variable file not passed")
 		tpl, err = readJobFile(string(src), flagVars)
 
 	default:
@@ -165,9 +158,4 @@ func readJobFile(src string, flagVars *map[string]string) (tpl *bytes.Buffer, er
 	}
 
 	return tpl, nil
-}
-
-// newTemplate returns an empty template with default options set
-func newTemplate() *template.Template {
-	return template.New("jobTemplate").Delims("[[", "]]").Option("missingkey=error")
 }
