+++ template/test-fixtures/single_templated.nomad
